#121 / 업데이트 로직 보수
