+++ app/src/main/java/sopt/uni/di/AppModule.kt
@@ -1,0 +1,21 @@
+package sopt.uni.di
+
+import android.content.Context
+import com.google.android.play.core.appupdate.AppUpdateManager
+import com.google.android.play.core.appupdate.AppUpdateManagerFactory
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object AppModule {
+
+    @Provides
+    @Singleton
+    fun provideAppUpdateManager(@ApplicationContext context: Context): AppUpdateManager =
+        AppUpdateManagerFactory.create(context)
+}
